+++ src/com/cloudogu/sos/pipebuildlib/dogu/stages/ReleaseStages.groovy
@@ -1,4 +1,4 @@
-package com.cloudogu.sos.pipebuildlib.stages
+package com.cloudogu.sos.pipebuildlib.dogu.stages
 
 import com.cloudogu.sos.pipebuildlib.*
 import com.cloudogu.sos.pipebuildlib.dogu.*
